Refactor: 알림 서비스와 관련된 패키지 정리
- NotificationPayload -> service.notification.payload
- NotificationService, SendNotificationCommand -> service.notification
- NotificationMapper, SendNotificationRequestMapper -> mapper.notification
