+++ src/main/java/com/zoopick/server/controller/NotificationController.java
@@ -3,9 +3,9 @@ package com.zoopick.server.controller;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.zoopick.server.dto.CommonResponse;
 import com.zoopick.server.dto.notification.*;
-import com.zoopick.server.mapper.SendNotificationRequestMapper;
+import com.zoopick.server.mapper.notification.SendNotificationRequestMapper;
 import com.zoopick.server.security.UserPrincipal;
-import com.zoopick.server.service.NotificationService;
+import com.zoopick.server.service.notification.NotificationService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
