+++ src/main/java/com/zoopick/server/service/ChatRoomService.java
@@ -10,8 +10,9 @@ import com.zoopick.server.repository.ChatMessageRepository;
 import com.zoopick.server.repository.ChatRoomRepository;
 import com.zoopick.server.repository.ItemRepository;
 import com.zoopick.server.repository.UserRepository;
-import com.zoopick.server.service.command.ChatMessagePayload;
-import com.zoopick.server.service.command.SendNotificationCommand;
+import com.zoopick.server.service.notification.NotificationService;
+import com.zoopick.server.service.notification.SendNotificationCommand;
+import com.zoopick.server.service.notification.payload.ChatMessagePayload;
 import lombok.RequiredArgsConstructor;
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
